Hack to prevent bot from remaining disconnected
Added a check that runs every 15 seconds, and stops the bot if it
hasn't received an RPC event in more than 15 seconds + the length of
the current song, or 30 minutes of inactivity when no song is playing.
Meant to be used with an auto-restarting process manager like PM2. The
option is enabled with the stopBotOnConnectionLoss config option.